refactor: js to ts (2024.05.03)
